requirements.txt
pyinstaller
python-vlc
sphinx
sphinx-rtd-theme

